Strip dollar sign from price when parsing single result
When a search returns a single result, there is an extra dollar sign in
the price:

    | 36-817 | Tektro V-brake lever MT20                     |   $16.00$ | Yes       |

This is because the library returns a price string with a dollar sign
when the search yields a single result, but without dollar sign when
the search yields multiple results.

The function `parse_info` already strips the dollar sign, so make
`parse_single_result` strip it as well, such that prices as consistently
returned in the `AA.BB` format.

With this change, the command-line displays the price correctly:

    | 36-817 | Tektro V-brake lever MT20                     |    16.00$ | Yes       |